Prevent glog from writing to /tmp during tests
glog (brought by client-go dep) registers an init() that configure the
default logger to write logs as files in /tmp. This is disabled by
sp13/pflags, which is loaded by default in Katafygio, but not during
tests.

So forcing glog to use stderr instead of files prevent a few files
from "leaking" in /tmp every time we run unit tests.