Use a temp file + rename for atomic changes
We were updating the files in place. This left a small time frame
where files may be commited while in the middle of a write. The
final, complete file would then be commited soon after, but we
would end up with an ugly, partial commit in the history.

Using a temporary file then a rename ensure git only see complete
files (rename is atomic on POSIX filesystems).