Katafygio logo
Copyright details:

This image or media file contains material based on a work of a National Park Service employee, created as part of that person's official duties. As a work of the U.S. federal government, such work is in the public domain in the United States. See the NPS website and NPS copyright policy for more information.

References:
https://www.nps.gov/aboutus/disclaimer.htm
https://thenounproject.com/term/sleeping-shelter/241/
https://commons.wikimedia.org/wiki/File:Pictograms-nps-accommodations-sleeping_shelter-2.svg
https://commons.wikimedia.org/wiki/File:Pictograms-nps-accommodations-sleeping_shelter.svg