Don't use tempdir
We shouldn't let Afero decide the temporary dir for us, since
this could end up on a different filesystem than the repository,
and rename() don't work cross filesystems.

We don't want to alter .gitignore (which can be edited/managed by
someone else or elsewhere), but we can use the local-only
.git/info/exclude exclusion file (https://git-scm.com/docs/gitignore).

Which allows us to create our temporary files in the repository
itself, and keep everything else (even /tmp) read-only.