Perform a shallow clone in the git provider for faster startup
Running katafygio for a couple of months on rather active cluster
results in a huge git repo.

On startup, katafygio clone the git repository, which in our case was
very slow, exceeding the configured deadlines.

This PR propose to perform a shallow clone with a depth of 1: katafygio
doesn't need to full history to work, and the startup is much snappier.